requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.4
numpy==1.23.3
Pillow==9.2.0
psycopg2-binary==2.9.3
requests==2.28.1
sqlparse==0.4.2
torch==1.12.1
torchaudio==0.12.1+cpu
torchvision==0.13.1
typing-extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
